Add files via upload
Have made a start on PHP code. Currently, the code will check to see if any POST data exists upon form submission. If the POST data exists, then there are several if statements which will check each form element to see if its empty or not. If the form elements, for example, name or email, are empty, then a PHP variable is created which will store a error message. The error messages are echoed next to the respective form elements upon form submission after being checked.