Add files via upload
Fixed 1 of 2 previous errors which was occurring. One of the errors was in regards to the preg_match() function and regular expression, however. a slight change has been made which has corrected this error. Now, if the user enters some data into form element, for example, the name input, the code will check if the data entered in the name input does not meet the specified format. If it doesn't, an error message is stored in a PHP variable which is printed out later. However, if the input is fine, no error message is printed. The PHP code for the other form elements needs to be updated.