Add files via upload
Updated PHP code for the email input. Now, there is a function which checks if the input entered for email matches a email format. If it does, no error message is printed. However, if an invalid email input is entered, an appropriate error message such as "Invalid email format, please enter the correct email format" is stored in a PHP variable and displayed next to the email input. As of now, the PHP input validation for name and email inputs appear to be working and appropriate error messages are displayed for incorrect inputs.