Add files via upload
Updated the names assigned to the form inputs. Also, now if a user enters an invalid input, for example, the email, after submitting the form, the POSTED value which the user entered will be displayed inside the original text box so that the user knows their input isn't right and need to try again.