Add files via upload
The error messages which are displayed (after POST data has been checked) next to invalid fields are now styled a different colour. Also, there is a counter included. With each input validation check, if the input entered is valid, no error message will be displayed. However, if the input entered is invalid, an error message will be displayed next to the input field and the "counter" will increment by 1.  At the end, there will be a final check, so if the value of the "counter" is equal to 0 then the POST data should be added to the SESSION. Otherwise (if the counter value is anything other than 0, e.g. 1,2,3) POST data is not added to SESSION and a message will be displayed on the screen to make this clear to the user.