Add files via upload
The receipt page displays business information (details about Lunardo) such as name, email etc as well as customer details. It's important to note that the customer details which are being displayed are not "hard coded". Instead, they are accessed from the SESSION data and will display according to what the user has entered. Also, there is a table which has been included on the receipt page. The table contains seat information such as quantity, unit cost and total cost. The quantity displayed depends on what the user enters, (NOT HARDCODED) and this also applies for calculation of the total cost. For all calculations involved, the numbers are displayed to "2 decimal places" with a "$" in front.