Add files via upload
Updated the receipt page by including a GST component within the table. The GST component is calculated by getting the subtotal and then dividing it by 11. Also, a "TOTAL" component has been added to the table. The TOTAL is calculated by adding the subtotal and the GST together. Note, just like all other calculations, the total is displayed to 2 decimal places and with a $ in front.