Add files via upload
Up-to-date receipt page which includes all the information such as business information, customer information and prices. Customer information is retrieved from the SESSION and not "hard coded". Regarding the prices, the seats are displayed in a table. The quantities and total prices displayed within the table are shown according to the user input (not hard coded). GST and Total price components are shown and calculated. Receipt page has been styled appropriately, keeping the receipt neat. Also, the PHP code to check if the SESSION is empty has been added. This is because if the user attempts to visit the receipt page and the SESSION is empty, then the user will be redirected to the index page.